♻️ Refactor: DataModel, UserStatus를 data/model 패키지로 이동
#17
