+++ app/src/main/java/com/example/saybettereducator/data/api/helper/FirebaseClient.kt
@@ -1,8 +1,8 @@
 package com.example.saybettereducator.data.api.helper
 
 import android.util.Log
-import com.example.saybettereducator.domain.model.DataModel
-import com.example.saybettereducator.domain.model.UserStatus
+import com.example.saybettereducator.data.model.DataModel
+import com.example.saybettereducator.data.model.UserStatus
 import com.example.saybettereducator.data.model.FirebaseFieldNames.LATEST_EVENT
 import com.example.saybettereducator.data.model.FirebaseFieldNames.STATUS
 import com.example.saybettereducator.utils.MyEventListener
