+++ app/src/main/java/com/example/saybettereducator/data/model/UserStatus.kt
@@ -1,4 +1,4 @@
-package com.example.saybettereducator.domain.model
+package com.example.saybettereducator.data.model
 
 enum class UserStatus {
     ONLINE, OFFLINE, IN_CALL
