+++ app/src/main/java/com/example/saybettereducator/data/repository/MainRepository.kt
@@ -1,8 +1,8 @@
 package com.example.saybettereducator.data.repository
 
-import com.example.saybettereducator.domain.model.DataModel
-import com.example.saybettereducator.domain.model.DataModelType.*
-import com.example.saybettereducator.domain.model.UserStatus
+import com.example.saybettereducator.data.model.DataModel
+import com.example.saybettereducator.data.model.DataModelType.*
+import com.example.saybettereducator.data.model.UserStatus
 import com.example.saybettereducator.data.api.helper.FirebaseClient
 import com.example.saybettereducator.utils.webrtcObserver.MyPeerObserver
 import com.example.saybettereducator.data.api.helper.WebRTCClient
